requirements.txt
confluent-kafka==2.5.0
python-decouple==3.8
smtplibaio==2.1.2.post1
